+++ packages/sphinx-autodoc-api-style/src/sphinx_autodoc_api_style/_transforms.py
@@ -14,7 +14,7 @@ from sphinx import addnodes
 from sphinx.util import logging as sphinx_logging
 
 from sphinx_autodoc_api_style._badges import build_badge_group
-from sphinx_autodoc_layout import inject_signature_slots
+from sphinx_ux_autodoc_layout import inject_signature_slots
 
 if t.TYPE_CHECKING:
     from sphinx.application import Sphinx