+++ packages/sphinx-autodoc-docutils/src/sphinx_autodoc_docutils/_directives.py
@@ -12,7 +12,7 @@ from sphinx import addnodes
 from sphinx.util.docutils import SphinxDirective
 
 from sphinx_autodoc_docutils._badges import build_kind_badge_group
-from sphinx_autodoc_layout import (
+from sphinx_ux_autodoc_layout import (
     ApiFactRow,
     build_api_facts_section,
     build_api_summary_section,