+++ packages/sphinx-autodoc-pytest-fixtures/src/sphinx_autodoc_pytest_fixtures/__init__.py
@@ -99,7 +99,7 @@ def setup(app: Sphinx) -> SetupDict:
     """
     app.setup_extension("sphinx.ext.autodoc")
     app.setup_extension("sphinx_ux_badges")
-    app.setup_extension("sphinx_autodoc_layout")
+    app.setup_extension("sphinx_ux_autodoc_layout")
     app.setup_extension("sphinx_autodoc_typehints_gp")
 
     import pathlib