+++ tests/ext/autodoc_docutils/test_doctree.py
@@ -12,7 +12,7 @@ from sphinx_autodoc_docutils._directives import (
     _normalize_directive_nodes,
     _normalize_role_nodes,
 )
-from sphinx_autodoc_layout._nodes import api_component
+from sphinx_ux_autodoc_layout._nodes import api_component
 
 
 class _DemoDirective(Directive):